+++ users/users-service/src/main/java/com/meysam/users/service/api/MemberService.java
@@ -4,7 +4,7 @@ import com.meysam.common.model.dto.*;
 import com.meysam.common.model.entity.Member;
 import org.springframework.http.ResponseEntity;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 
 public interface MemberService/* extends UserDetailsService*/ {
@@ -12,7 +12,7 @@ public interface MemberService/* extends UserDetailsService*/ {
   ResponseEntity<LoginResponseDto> login(LoginRequestDto loginRequestDto);
   ResponseEntity register(RegisterUserRequestDto registerUserRequestDto);
   Member findByUserName(String username);
-  Member findById(BigDecimal id);
+  Member findById(BigInteger id);
   List<Member> findAll();
   UserWalletsDto getUserWallets(String token,String username);
   ResponseEntity<String> createMemberWalletAddress(String token,MemberWalletDto memberWalletDto);
