+++ users/users-service/src/main/java/com/meysam/users/service/impl/MemberServiceImpl.java
@@ -14,7 +14,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 
 @Slf4j
@@ -49,7 +49,7 @@ public class MemberServiceImpl implements MemberService {
     }
 
     @Override
-    public Member findById(BigDecimal id) {
+    public Member findById(BigInteger id) {
         return memberRepository.findById(id).orElse(null);
     }
